hotfix add validation for start end date
