+++ Server/src/main/java/lt/techin/eventify/model/UserImage.java
@@ -6,67 +6,67 @@ import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name =  "user_images")
+@Table(name = "user_images")
 public class UserImage {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    private String filename;
-    private String contentType;
+  private String filename;
+  private String contentType;
 
-    @Lob
-    @Column(name = "data", columnDefinition = "LONGBLOB")
-    private byte[] data;
+  @Lob
+  @Column(name = "data", columnDefinition = "LONGBLOB")
+  private byte[] data;
 
-    private Long fileSize;
-    private LocalDateTime uploadedAt;
+  private Long fileSize;
+  private LocalDateTime uploadedAt;
 
-    public UserImage() {
+  public UserImage() {
 
-    }
+  }
 
-    public Long getId() {
-        return id;
-    }
+  public Long getId() {
+    return id;
+  }
 
-    public String getFilename() {
-        return filename;
-    }
+  public String getFilename() {
+    return filename;
+  }
 
-    public void setFilename(String filename) {
-        this.filename = filename;
-    }
+  public void setFilename(String filename) {
+    this.filename = filename;
+  }
 
-    public String getContentType() {
-        return contentType;
-    }
+  public String getContentType() {
+    return contentType;
+  }
 
-    public void setContentType(String contentType) {
-        this.contentType = contentType;
-    }
+  public void setContentType(String contentType) {
+    this.contentType = contentType;
+  }
 
-    public byte[] getData() {
-        return data;
-    }
+  public byte[] getData() {
+    return data;
+  }
 
-    public void setData(byte[] data) {
-        this.data = data;
-    }
+  public void setData(byte[] data) {
+    this.data = data;
+  }
 
-    public Long getFileSize() {
-        return fileSize;
-    }
+  public Long getFileSize() {
+    return fileSize;
+  }
 
-    public void setFileSize(Long fileSize) {
-        this.fileSize = fileSize;
-    }
+  public void setFileSize(Long fileSize) {
+    this.fileSize = fileSize;
+  }
 
-    public LocalDateTime getUploadedAt() {
-        return uploadedAt;
-    }
+  public LocalDateTime getUploadedAt() {
+    return uploadedAt;
+  }
 
-    public void setUploadedAt(LocalDateTime uploadedAt) {
-        this.uploadedAt = uploadedAt;
-    }
+  public void setUploadedAt(LocalDateTime uploadedAt) {
+    this.uploadedAt = uploadedAt;
+  }
 }
