+++ Server/src/main/java/lt/techin/eventify/validation/ConsistentAgeValidation/ConsistentAgeRange.java
@@ -1,0 +1,4 @@
+package lt.techin.eventify.validation.ConsistentAgeValidation;
+
+public @interface ConsistentAgeRange {
+}
