--- Server/src/main/java/lt/techin/eventify/validation/ValidImage.java
@@ -1,16 +1,0 @@
-package lt.techin.eventify.validation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-@Documented
-@Constraint(validatedBy = ImageFileValidator.class)
-@Target({ElementType.FIELD,ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ValidImage {
-    String message() default "Invalid image file";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
